// src/app/page/master/agama/agama.interface.ts
import {EventEmitter} from '@angular/core';

export interface Agama {
	kdProfile?: string;
	kdAgama?: string;
	namaAgama?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;

 }